+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.0 (June 13th, 2013)
 
+- Add possibility to fetch OrgId after login
 - Update default Salesforce API version to 28.0
 
 ## 1.0.0 (May 13th, 2013)